requirements.txt
requests
beautifulsoup4
easyocr
pillow
opencv-python-headless
PyMuPDF
urllib3
